// src/hoat-dong-day-hoc/interfaces/hoat-dong-day-hoc.interface.ts
export interface IHoatDongDayHoc {
  readonly id?: number;
  readonly idCD?: number;
  readonly ma?: string;
  readonly ten?: string;
  readonly stt?: number;
  readonly createdAt?: Date;
  readonly updatedAt?: Date;
  readonly createdBy?: number;
  readonly updatedBy?: number;
  readonly isDeleted?: boolean;
}